requirements.txt
async-timeout==4.0.2
fastapi==0.92.0
pydantic==1.10.5
python-dateutil==2.8.2
starlette==0.25.0
typing_extensions==4.4.0
uvicorn==0.19.0
websockets==10.4
kedro==0.18.4
python-dotenv==0.19.0
pandas==1.3.3
